requirements.txt
streamlit
requests
beautifulsoup4
lxml
faiss-cpu
python-dotenv
numpy
scikit-learn
openai==0.28.1
google-generativeai
